pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[project]
name = "hera_sim"
description = "A collection of simulation routines describing the HERA instrument."
readme = "README.rst"
authors = [
  {name="HERA Team"},
  {name="Steven Murray", email="murray.steveng@gmail.com"}
]
license = {text="MIT"}
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Scientific/Engineering :: Astronomy",
  "Topic :: Scientific/Engineering :: Physics",
]
dynamic=["version"]
requires-python = ">=3.10"

dependencies = [
  "astropy",
  "astropy-healpix",
  "cached-property",
  "deprecation",
  "hera-cli-utils>=0.1.0",
  "numpy>=2",
  "pyuvdata>=3.1.2",
  "pyuvsim>=1.4",
  "pyyaml>=5.1",
  "rich",
  "scipy",
  "typing-extensions;python_version<'3.11'",
]

[project.optional-dependencies]
all = [
  "hera-sim[vis,bda,cal,gpu]",
]
bda = [
  "bda"
]
cal = [
  "hera-calibration>3.6.1"
]
dev = [
  "hera-sim[docs,tests]",
]
docs = [
  "fftvis",
  "furo",
  "ipython",
  "matvis>=1.3.0",
  "nbsphinx",
  "numpydoc>=0.8",
  "pyradiosky>=0.1.2",
  "sphinx>=1.8,<7.2",
  "sphinx-autorun",
]
tests = [
  "coverage>=4.5.1",
  "hera-sim[bda,cal,vis]",
  "matplotlib>=3.4.2",
  "pre-commit",
  "pytest>=3.5.1",
  "pytest-cov>=2.5.1",
  "uvtools",
]
vis = [
  "fftvis>=0.1.0",
  "line-profiler",
  "matvis>=1.2.1",
  "mpi4py",
  "pyradiosky>=0.1.2",
]

[project.urls]
Documentation = "https://hera_sim.readthedocs.org"
Repository = "https://github.com/HERA-Team/hera_sim"

[tool.setuptools]
script-files = ["scripts/hera-sim-simulate.py", "scripts/hera-sim-vis.py"]

[tool.pytest.ini_options]
addopts = "--cov hera_sim --cov-config=.coveragerc --cov-report xml:./coverage.xml --durations=25 --verbose"
norecursedirs = [
  "dist",
  "build",
  ".tox",
]
testpaths = "hera_sim/tests"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = [
  "UP",   # pyupgrade
  "E",    # pycodestyle
  "W",    # pycodestyle warning
  "NPY",  # numpy-specific rules
  "I",    # isort
]

ignore = [
  "NPY002",  # RNG -- fix soon!
]
